Added Tutorial Scene and level Textures
Added a small Tutorial scene still need to tune it, to make it feel right and also add text telling the player how to play the game. Didn't check if the time is affected by this but boundaries are also fix for the entire game. Please don't switch the tags